separate template from container, add onToggle prop to dropdown
