+++ src/modules/elements/search-posts/style.pcss
@@ -19,7 +19,7 @@
 		margin: 0;
 	}
 
-	.tribe-editor__search-posts__results-list-item-action {
+	.tribe-editor__search-posts__results-list-item-button {
 		display: block;
 		width: 100%;
 		color: #555d66;
